index.htmlのみcookieのuser発行をするようにする
